Reorganize directory and module structure
