+++ benchmarker/__init__.py
@@ -1,0 +1,2 @@
+from .tester import TestData, Tester, TestResult
+from .satsolver import SatSolver
