+++ data/config.json
@@ -1,0 +1,21 @@
+{
+    "resultsDir": "benchmarks/",
+    "puzzleDir": "data/puzzles/",
+    "cacheDir": ".cache/",
+    "round": 2,
+    "defaultPuzzleSet": "Standard",
+    "puzzleSets": {
+        "Standard": {
+            "file": "p096_sudoku.txt",
+            "numPuzzles": 50,
+            "size": 9,
+            "offset": 1
+        },
+        "Hard": {
+            "file": "top95",
+            "numPuzzles": 95,
+            "size": 1,
+            "offset": 0
+        }
+    }
+}
