+++ satcoder/__init__.py
@@ -1,0 +1,2 @@
+from satcoder.cnf import Encoding, encode
+from satcoder.sudoku import decode
